Moved peek method back to SgfStream
The peek_at_next_non_whitespace method had originally looked like it
belonged on SgfStream, and it still does. Now that it doesn't have any
of the still_inside business attached to it, we can move it back.
Also renamed it to peek_skipping_whitespace, since that seems more
concise and expressive.