Delete outdated comment.
The fields and where they're used reference is no longer up-to-date,
and is no longer that useful. The next step is going to be tricky:
Parser is still doing a lot, but it seems risky to have too many
different classes that all deal with the SgfStream -- that'll just make
the logic harder to follow. It's already a bit iffy spreading across
all those different tokens. The best hope as I see it: extract all the
SgfStream-using stuff out from the middle of Parser, creating a class
that wraps an SgfStream and presents Parser with a higher-level
abstraction.